fix(QScrollArea): prevent content re-rendering on scroll or mousemove
